# packages/contracts/openapi/cms.yaml
openapi: 3.1.0
info:
  title: Candidate Supply CMS API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /work-items:
    post:
      operationId: createWorkItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkItemRequest'
      responses:
        '201':
          description: Created manual work item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiWorkItemEnvelope'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
          csrfHeader: []
    get:
      operationId: listWorkItems
      parameters:
        - in: query
          name: view
          required: false
          schema:
            type: string
            default: actionable
        - in: query
          name: sort
          required: false
          schema:
            type: string
            default: priority
        - in: query
          name: query
          required: false
          schema:
            type: string
        - in: query
          name: cursor
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Work queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiWorkItemsResponseEnvelope'
        default:
          $ref: '#/components/responses/Unauthorized'
  /work-items/summary:
    get:
      operationId: getWorkSummary
      parameters:
        - in: query
          name: view
          required: false
          schema:
            type: string
            default: actionable
      responses:
        '200':
          description: Work queue KPI summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiWorkSummaryResponseEnvelope'
        default:
          $ref: '#/components/responses/Unauthorized'
  /work-items/{id}:
    get:
      operationId: getWorkItem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Work item detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiWorkItemEnvelope'
        default:
          $ref: '#/components/responses/Unauthorized'
    patch:
      operationId: updateWorkItem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkItemUpdate'
      responses:
        '200':
          description: Updated work item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiWorkItemEnvelope'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
          csrfHeader: []
...
  /tasks:
    get:
      operationId: listTasks
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [NEW, IN_PROGRESS, DONE, CANCELLED]
        - in: query
          name: assigneeUserId
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Scoped task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiTasksResponseEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
    post:
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiTaskEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
          csrfHeader: []
  /tasks/{id}:
    get:
      operationId: getTask
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Scoped task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiTaskEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
  /tasks/{id}/assignment:
    post:
      operationId: assignTask
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignTaskRequest'
      responses:
        '200':
          description: Task assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiTaskEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
          csrfHeader: []
  /tasks/{id}/start:
    post:
      operationId: startTask
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedTaskActionRequest'
      responses:
        '200':
          description: Task started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiTaskEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
          csrfHeader: []
  /tasks/{id}/completion:
    post:
      operationId: completeTask
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedTaskActionRequest'
      responses:
        '200':
          description: Task completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiTaskEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
          csrfHeader: []
  /tasks/{id}/cancellation:
    post:
      operationId: cancelTask
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedTaskActionRequest'
      responses:
        '200':
          description: Task cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiTaskEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
          csrfHeader: []
...
components:
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    Conflict:
      description: Optimistic concurrency or duplicate relation conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    ValidationError:
      description: Business validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    Forbidden:
      description: Permission required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
...
  schemas:
...
    ApiProblem:
      type: object
      required:
        - error
        - requestId
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
        requestId:
          type: string
          minLength: 1
...
    PersonRef:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    CandidateRef:
      type: object
      required:
        - id
        - code
        - name
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
    ClientRef:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    OrderRef:
      type: object
      required:
        - id
        - code
        - position
      properties:
        id:
          type: string
        code:
          type: string
        position:
          type: string
    WorkItem:
      type: object
      required:
        - id
        - title
        - priority
        - status
        - dueAt
        - assignee
        - sourceType
        - sourceLabel
        - candidate
        - order
        - client
        - updatedAt
        - version
        - waitingOn
      properties:
        id:
          type: string
        title:
          type: string
        priority:
          type: string
          enum:
            - URGENT
            - HIGH
            - NORMAL
        status:
          type: string
          enum:
            - NEW
            - IN_PROGRESS
            - DONE
            - CANCELLED
        dueAt:
          type: string
          format: date-time
        assignee:
          $ref: '#/components/schemas/PersonRef'
        sourceType:
          type: string
        sourceLabel:
          type: string
        candidate:
          $ref: '#/components/schemas/CandidateRef'
        order:
          $ref: '#/components/schemas/OrderRef'
        client:
          $ref: '#/components/schemas/ClientRef'
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          minimum: 1
        lastActivity:
          type: string
        notes:
          type: string
        waitingOn:
          type: string
          enum:
            - NONE
            - CANDIDATE
            - CLIENT_PARTNER
            - INTERNAL
            - SYSTEM
    WorkItemsResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkItem'
    WorkSummary:
      type: object
      required:
        - overdue
        - today
        - waitingReply
        - unresolvedEmail
        - journeyRisk
      properties:
        overdue:
          type: integer
        today:
          type: integer
        waitingReply:
          type: integer
        unresolvedEmail:
          type: integer
        journeyRisk:
          type: integer
    WorkSummaryResponse:
      type: object
      required:
        - summary
      properties:
        summary:
          $ref: '#/components/schemas/WorkSummary'
    WorkItemUpdate:
      type: object
      required:
        - version
      properties:
        status:
          type: string
          enum:
            - NEW
            - IN_PROGRESS
            - DONE
            - CANCELLED
        dueAt:
          type: string
          format: date-time
        assigneeId:
          type: string
        reason:
          type: string
        version:
          type: integer
          minimum: 1
        waitingOn:
          type: string
          enum:
            - NONE
            - CANDIDATE
            - CLIENT_PARTNER
            - INTERNAL
            - SYSTEM
...
    CreateWorkItemRequest:
      type: object
      required:
        - title
        - priority
        - dueAt
        - candidateId
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 240
        priority:
          type: string
          enum:
            - URGENT
            - HIGH
            - NORMAL
        dueAt:
          type: string
          format: date-time
        candidateId:
          type: string
        orderId:
          type: string
          nullable: true
        clientId:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
          maxLength: 2000
...
    Task:
      type: object
      required:
        - id
        - title
        - status
        - assigneeUserId
        - teamId
        - waitingOn
        - dueAt
        - noDueDateReason
        - ruleCode
        - sourceEventId
        - dedupeKey
        - referenceEntityType
        - referenceEntityId
        - version
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [NEW, IN_PROGRESS, DONE, CANCELLED]
        assigneeUserId:
          type: string
        teamId:
          type: string
          nullable: true
        waitingOn:
          type: string
          enum: [CANDIDATE, CLIENT_PARTNER, INTERNAL, OTHER]
          nullable: true
        dueAt:
          type: string
          format: date-time
          nullable: true
        noDueDateReason:
          type: string
          nullable: true
        ruleCode:
          type: string
          nullable: true
        sourceEventId:
          type: string
          nullable: true
        dedupeKey:
          type: string
          nullable: true
        referenceEntityType:
          type: string
        referenceEntityId:
          type: string
        version:
          type: integer
          minimum: 1
    CreateTaskRequest:
      type: object
      required: [title, assigneeUserId, referenceEntityType, referenceEntityId]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 240
        description:
          type: string
          nullable: true
        assigneeUserId:
          type: string
        teamId:
          type: string
          nullable: true
        waitingOn:
          type: string
          enum: [CANDIDATE, CLIENT_PARTNER, INTERNAL, OTHER]
          nullable: true
        dueAt:
          type: string
          format: date-time
          nullable: true
        noDueDateReason:
          type: string
          nullable: true
        ruleCode:
          type: string
          nullable: true
        sourceEventId:
          type: string
          nullable: true
        dedupeKey:
          type: string
          nullable: true
        referenceEntityType:
          type: string
        referenceEntityId:
          type: string
    VersionedTaskActionRequest:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          minimum: 1
        reason:
          type: string
          nullable: true
    AssignTaskRequest:
      type: object
      required: [version, assigneeUserId]
      properties:
        version:
          type: integer
          minimum: 1
        assigneeUserId:
          type: string
        teamId:
          type: string
          nullable: true
    TasksResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    ApiTaskEnvelope:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/Task'
        requestId:
          type: string
    ApiTasksResponseEnvelope:
      type: object
      required: [data, requestId]
      properties:
        data:
          $ref: '#/components/schemas/TasksResponse'
        requestId:
          type: string
...
    ApiPage:
      type: object
      required:
        - nextCursor
        - hasMore
      properties:
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
    ApiError:
      type: object
      required:
        - code
        - messageKey
      properties:
        code:
          type: string
        messageKey:
          type: string
        params:
          type: object
          additionalProperties: true
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        currentVersion:
          type: integer
          minimum: 0
          nullable: true
...
    ApiWorkItemEnvelope:
      type: object
      required:
        - data
        - requestId
      properties:
        data:
          $ref: '#/components/schemas/WorkItem'
        requestId:
          type: string
          minLength: 1
    ApiWorkItemsResponseEnvelope:
      type: object
      required:
        - data
        - page
        - requestId
      properties:
        data:
          $ref: '#/components/schemas/WorkItemsResponse'
        page:
          $ref: '#/components/schemas/ApiPage'
        requestId:
          type: string
          minLength: 1
    ApiWorkSummaryResponseEnvelope:
      type: object
      required:
        - data
        - requestId
      properties:
        data:
          $ref: '#/components/schemas/WorkSummaryResponse'
        requestId:
          type: string
          minLength: 1
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cms_sid
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token